requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
bandit==1.7.5
